refactor: add useCreateProductFetch.ts and use in CreateProductModal.tsx
