+++ src/components/CreateProductModal.tsx
@@ -1,9 +1,6 @@
-import NumberFormatter from "@/lib/NumberFormatter";
+import { useCreateProductFetch } from "@/hooks/useCreateProductFetch";
 import ModalInput from "./ModalInput";
-import { useState } from "react";
-import { PublicError } from "@/lib/CustomErrors";
 import { CgSpinner } from "react-icons/cg";
-import { useSWRConfig } from "swr";
 
 interface ModalProps {
   isModalOpen: boolean;
@@ -11,64 +8,8 @@ interface ModalProps {
   userId: string;
 }
 
-interface INewProduct {
-  productName: string;
-  productPrice: string;
-  productQuantity: string;
-}
-
 export default function CreateProductModal({ isModalOpen, userId, onModalOpen }: ModalProps) {
-  const [errorMessage, setErrorMessage] = useState("");
-  const [errorAction, setErrorAction] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const [product, setProduct] = useState<INewProduct>({
-    productName: "",
-    productPrice: "",
-    productQuantity: "",
-  });
-
-  const { mutate } = useSWRConfig();
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    setErrorMessage("");
-    setErrorAction("");
-    setIsLoading(true);
-
-    try {
-      const response = await fetch(`/api/v1/product/create-product/${userId}`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          productName: product.productName,
-          productPrice: NumberFormatter.toNumber(product.productPrice),
-          productQuantity: NumberFormatter.toNumber(product.productQuantity),
-        }),
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json();
-        if (errorData.isPublicError) {
-          throw new PublicError(errorData.message, errorData.action);
-        }
-      }
-
-      mutate(`/api/v1/product/list-products/${userId}`);
-      onModalOpen(false);
-      setIsLoading(false);
-      setProduct({ productName: "", productPrice: "", productQuantity: "" });
-    } catch (error) {
-      if (error instanceof PublicError) {
-        setErrorMessage(error.message);
-        setErrorAction(error.action);
-        setIsLoading(false);
-        setTimeout(() => {
-          setErrorMessage("");
-          setErrorAction("");
-        }, 2000);
-      }
-    }
-  }
+  const { product, setProduct, isLoading, errorMessage, errorAction, handleSubmit } = useCreateProductFetch();
 
   return (
     <section
@@ -81,7 +22,7 @@ export default function CreateProductModal({ isModalOpen, userId, onModalOpen }:
 
         <h2 className="text-lg mb-6 text-lightTxt dark:text-darkTxt">Adicione um novo produto</h2>
 
-        <form className="flex flex-col gap-5" onSubmit={(e) => handleSubmit(e)}>
+        <form className="flex flex-col gap-5" onSubmit={(e) => handleSubmit(e, userId)}>
           <ModalInput
             type="text"
             placeholder="Nome"
